requirements.txt
atlassian-python-api==3.41.3
configparser==6.0.0
openpyxl==3.1.2
requests==2.31.0